automation/requirements.txt
google-genai
google-generativeai
python-dotenv
feedparser
python-dateutil
markdown
google-cloud-aiplatform
Pillow
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0
tweepy